+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Add the project name in the favourites list
+- Refresh favourites list on toggle one favourite
 
 ### Fixed