+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Add the project name in the favourites list
 - Refresh favourites list on toggle one favourite
+- Add a default message if there are not tasks to send
 
 ### Fixed